Renames Arrow to Chevron and adds variations up/left/right
